backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
cohere>=4.9.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytest>=7.0.0